requirements.txt
playwright~=1.53.0
allure-python-commons~=2.14.3
pytest~=8.4.1